mm/ksm: always merge anonymous memory
By default, KSM works only on memory that is added by madvise(). And
the only way to get around that is to either:
  * use LD_PRELOAD; or
  * patch the kernel with UKSM or PKSM.

Instead, lets implement a so-called "always" mode, which allows marking
VMAs as mergeable on do_anonymous_page() call.

The patch introduces a new sysctl knob as well as kernel cmdline option
to control which mode to use. The default mode is to maintain old
(madvise-based) behaviour.

TODO:

  * add VM_UNMERGEABLE for opt-out due to security reasons;
  * re-scan all the VMAs after switching to the "always" mode (using
  workqueue?);
  * statistics.

Signed-off-by: Oleksandr Natalenko <oleksandr@natalenko.name>