lrng: limit kernel crypto API default DRNGs
When the LRNG is compiled as the /dev/random provided (i.e. the legacy RNG
is not compiled by not selecting CONFIG_RANDOM_DEFAULT_IMPL), only the
ChaCha20 DRNG can be used as the default DRNG. This is due to the fact
that the LRNG must be capable of delivering random numbers very early
in the boot cycle. At that time, it is not guaranteed that the kernel
crypto API is available.

To select a kernel crypto API DRNG to be automatically loaded during boot,
select CONFIG_LRNG_SWITCH_DRBG=y. This selection allocates, for example,
the SP800-90A DRBG during boot, but at the time the kernel crypto API is
present.

To enforce this, CONFIG_LRNG_DFLT_DRNG_DRBG and CONFIG_LRNG_DFLT_DRNG_KCAPI
is only present if the legacy RNG is compiled and thus the LRNG does not
provide random numbers during early boot time.

Signed-off-by: Stephan Mueller <smueller@chronox.de>